update e delete expenses

+++ back/controllers/ExpenseController.js
@@ -45,14 +45,57 @@ module.exports = {
     },
 
     async update(req, res) {
-        const expense = await Expense.findByIdAndUpdate(req.params.expID, req.body, { new: true });
+        var token = req.headers['x-access-token'];
+        if (!token) return res.status(401).send({ user: {}, message: "Nenhum token foi fornecido." });
+        
+        jwt.verify(token, authConfig.secret, function(err, decoded) {
+        
+            if (err) return res.status(500).send({ user: {}, message: "Falha ao autenticar token." });
 
-        return res.json(expense);
+          User.findById(decoded.id, 
+                { password: 0 }, 
+                async function (err, user) {
+                    if (err) return res.status(500).send("Houve um problema ao encontrar o usuario");
+                    if (!user) return res.status(404).send("Nenhum usuário encontrado.");                
+                    
+                    const expense = await Expense.findByIdAndUpdate(req.params.expID, req.body, { new: true });
+                    return res.json(expense);
+                        
+                }
+            )
+        })
+        
+
     },
 
     async delete(req,res){
-        const expense = await Expense.findByIdAndDelete(req.params.expID);
+        var token = req.headers['x-access-token'];
+        if (!token) return res.status(401).send({ user: {}, message: "Nenhum token foi fornecido." });
+        
+        jwt.verify(token, authConfig.secret, function(err, decoded) {
+        
+            if (err) return res.status(500).send({ user: {}, message: "Falha ao autenticar token." });
+
+
+          User.findById(decoded.id, 
+                { password: 0 }, 
+                async function (err, user) {
+                    if (err) return res.status(500).send("Houve um problema ao encontrar o usuario");
+                    if (!user) return res.status(404).send("Nenhum usuário encontrado.");                
+                    
+                    const expense = await Expense.findByIdAndDelete(req.params.expID);
+                    return res.json(expense);
+                        
+                }
+            )
+        })
+        
+
+    }
+
+    };
+/*
+ const expense = await Expense.findByIdAndUpdate(req.params.expID, req.body, { new: true });
 
         return res.json(expense);
-    }
-};
+*/
